+++ crates/usvg-tree/src/lib.rs
@@ -850,18 +850,6 @@ pub struct Path {
     /// `shape-rendering` in SVG.
     pub rendering_mode: ShapeRendering,
 
-    /// Contains a text bbox.
-    ///
-    /// Text bbox is different from path bbox. The later one contains a tight path bbox,
-    /// while the text bbox is based on the actual font metrics and usually larger than tight bbox.
-    ///
-    /// Also, path bbox doesn't include leading and trailing whitespaces,
-    /// because there is nothing to include. But text bbox does.
-    ///
-    /// As the name suggests, this property will be set only for paths
-    /// that were converted from text.
-    pub text_bbox: Option<NonZeroRect>,
-
     /// Segments list.
     ///
     /// All segments are in absolute coordinates.
@@ -878,7 +866,6 @@ impl Path {
             stroke: None,
             paint_order: PaintOrder::default(),
             rendering_mode: ShapeRendering::default(),
-            text_bbox: None,
             data,
         }
     }
@@ -1318,6 +1305,12 @@ fn calc_node_bbox(node: &Node, ts: Transform) -> Option<BBox> {
 
             Some(bbox)
         }
-        NodeKind::Text(_) => None,
+        NodeKind::Text(ref text) => {
+            if let Some(bbox) = text.bounding_box {
+                bbox.transform(ts).map(BBox::from)
+            } else {
+                None
+            }
+        }
     }
 }