+++ crates/usvg-tree/src/text.rs
@@ -7,6 +7,7 @@ use std::rc::Rc;
 use strict_num::NonZeroPositiveF32;
 
 use crate::{Fill, Node, PaintOrder, Stroke, TextRendering, Visibility};
+use tiny_skia_path::NonZeroRect;
 
 /// A font stretch property.
 #[allow(missing_docs)]
@@ -322,6 +323,17 @@ pub struct Text {
 
     /// A list of text chunks.
     pub chunks: Vec<TextChunk>,
+
+    /// Contains a text bounding box.
+    ///
+    /// Text bounding box is special in SVG and doesn't represent
+    /// a tight bounding box of the element's content.
+    /// You can find more about it [here](https://razrfalcon.github.io/notes-on-svg-parsing/text/bbox.html).
+    ///
+    /// Will be set by
+    /// [`usvg_text_layout::TreeTextToPath::convert_text`](
+    /// https://docs.rs/usvg-text-layout/latest/usvg_text_layout/trait.TreeTextToPath.html#tymethod.convert_text)
+    pub bounding_box: Option<NonZeroRect>,
 
     /// Text converted into paths, ready to render.
     ///